requirements.txt
tensorflow==2.3.0
scikit-learn==0.22.1
matplotlib
python-dotenv
azureml-sdk
azure-cli
